requirements.txt
crewai>=0.28.0
pandas>=2.0.0
pyyaml>=6.0
langchain>=0.0.267
python-dotenv>=1.0.0